[front] Merge new rent modal form
